Android: Upgrade to RN 0.71.0-rc.6
