+++ example/babel.config.js
@@ -1,7 +1,4 @@
 module.exports = {
   presets: ["module:metro-react-native-babel-preset"],
-  plugins: [
-    "react-native-reanimated/plugin",
-    "transform-inline-environment-variables",
-  ],
+  plugins: ["transform-inline-environment-variables"],
 };
